Fixing the dependency tree and neodev.nvim
mason-lspconfig.nvim is a bridge and depends on both
mason.nvim and nvim-lspconfig. So it must be higher
in the dependency tree.

neodev.nvim must be loaded before nvim-lspconfig loads.

checkThirdParty must be disabled for fixing a pop-up
that annoys the user:
" Do you need to configure your work environment as `luv`? "